Lots of cleanup and making stuff work
The current state of the component creates a sensor with the current
energy price (Currency per energy unit) as state, as well as the prices
for the current, and (if available) next day.

There is also some simple statistics, i.e. peak price and time, min
price and time, and average price.